test for number 5 on checkpoint 4 (#29)
This test is testing the functionality of the classes i made for earlier on in this checkpoint, it writes the numbers for the collatz sequence into the txt file, which then gets processed and returned by the usercomputeapiimpl java file. This test correctly outputted each number in the txt file, along with their collatz sequences.